backend/requirements.txt
allyin==0.1.4
fastapi==0.116.1
langchain==0.3.26
pydantic==2.11.7
PyPDF2==3.0.1
python-dotenv==1.1.1
qdrant_client==1.14.3
requests>=2.28.1,<3
# sentence_transformers==5.0.0
sentence_transformers[cpu]
uvicorn==0.35.0
apache-airflow==3.0.3
fpdf==1.7.2
pandas==2.3.0
python_docx==1.1.2
cryptography==41.0.7
bcrypt==4.1.2
xlsxwriter==3.1.9
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
reportlab==4.0.7
pyautogen==0.9.1a1 
praw==7.7.1
linkedin-api==2.2.0
beautifulsoup4==4.13.4
firecrawl-py==0.0.16
crawl4ai
redis==5.0.1